Proof of concept - automated accessibility testing
This proof of concept looks at automated accessibility testing and possible
approaches that would allow it to be added to our CI/CD pipelines.

The first approach involves setting up rake tasks that execute against paths
of the site using htmlhint for html validation and pa11y for accessibility
testing. It should be fairly easy to add a rake task to the pipeline as a
step so this is a promising approach.

The second approach involves adding integration tests using axe core in
selenium. An example test for this has been written, but these type of tests
could be written as part of our existing suite and executed during the existing
test step of the pipeline.

Both approaches could also be used together if desired.