WIP - adding schemes question to setup
Freezing work on this currently as we're refactoring how the setup section
is handled and most of the work on this ticket will be within the setup
section. Doing this now would invovle having to redo work once that setup is
complete so just adding this WIP in the meantime to return to this after
the refactoring is done.