Extract base log and create sales log (#858)
* Add abstract log class and sales log class

Created a parent log class for sales log and lettings log. Any bits common
to both sales and lettings can live in the parent class. As the sales log
functionality is built up any commonalities with lettings log can be extracted
into the parent log class. The sales log model is set up without a json form
and instead the form is defined in code - like the setup section of the lettings
log.

* update sales logs controller

* update lettings controller specs

* update filter method name

* update organisations controller

* use lettings method

* Add deleted tests back

* lint

Co-authored-by: Kat <katrina@madetech.com>
Co-authored-by: Kat <kosiak.katrina@gmail.com>